removed user_test.go moved to pinot_test
